+++ tests/test_boolean_arithmetics.py
@@ -2,7 +2,6 @@
 # SPDX-License-Identifier: BSD-3-Clause
 
 import operator as ops
-import sys
 
 import numpy as np
 import pytest
@@ -51,8 +50,6 @@ def test_bool_arithmetics_with_python_scalar(op, dtype, reflected):
             return op(b, a)
         return op(a, b)
 
-    strict = not (sys.platform.startswith("win") and np.__version__ < "2")
-
     np.testing.assert_array_equal(
-        do(ndx, reflected).unwrap_numpy(), do(np, reflected), strict=strict
+        do(ndx, reflected).unwrap_numpy(), do(np, reflected), strict=True
     )