+++ tests/test_searching_functions.py
@@ -19,9 +19,7 @@ def test_count_nonzero_bool_and_string(values):
         arr = npx.asarray(values)
         return npx.count_nonzero(arr)
 
-    np.testing.assert_array_equal(
-        do(ndx).unwrap_numpy(), do(np), strict=np.__version__ > "2"
-    )
+    np.testing.assert_array_equal(do(ndx).unwrap_numpy(), do(np), strict=True)
 
 
 @pytest.mark.parametrize(
@@ -37,6 +35,4 @@ def test_nonzero_bool_and_string(values):
         return npx.nonzero(arr)
 
     for ndx_res, np_res in zip(do(ndx), do(np), strict=True):
-        np.testing.assert_array_equal(
-            ndx_res.unwrap_numpy(), np_res, strict=np.__version__ > "2"
-        )
+        np.testing.assert_array_equal(ndx_res.unwrap_numpy(), np_res, strict=True)